fix defining pf particle tower producer twice
move producer to plugins
